[ZV-427]: count filtered (#111)
* [ZV-400] : implement order get, init testing configs

* [ZV-400] : add data to other services for test

* [ZV-400] : new station mapping

* [ZV-400] : new station mapping and dockerfile

* [ZV-400] : bump gradle ver

* [ZV-400] : refac

* Update docker-compose.yaml


* [ZV-400] : refac

* [ZV-401] : get waterfall

* [ZV-401] : refac, tests, get waterfall no sort, no filter

* [ZV-401] : redundant plugin

* [ZV-420] : sorting, filtering and seek

* [ZV-420] : refac + tests

* [ZV-420] : refac

* [ZV-420] : remove redundant r2dbc config

* [ZV-425] : get customer endpt

* [ZV-426] : create order endpt

* [ZV-425] : refac after merge

* [ZV-426] : refac after merge

* [ZV-426] : refac after merge

* [ZV-427] : filtered count

* [ZV-427] : bump library ver

* [ZV-427] : remove duplicate class from merge

---------

+++ order-service/src/test/kotlin/ru/zveron/order/test/util/TestData.kt
@@ -121,7 +121,7 @@ fun testProfile() = Profile(
 fun testCreateOrderEntrypointRequest() = createOrderRequest {
     this.animalId = randomId()
     this.profileId = randomId()
-    this.subwayStationId = randomSubwayId().toLong()
+    this.subwayStationId = randomSubwayId()
 
     this.price = RandomUtils.nextLong()
     this.title = RandomStringUtils.randomAlphabetic(10)
